# Turn device screen off _after_ everything has been set up