Sometimes it can take quite a while for everything to get set up and the screen to appear.

Note:
Intentionally kept the double `if` construct so it matches the previous blocks.