@rom1v 2023-04-12
> Intentionally kept the double if construct so it matches the previous blocks.

I merged the conditions in a single `if`, rebased on `dev`, and adapted the commit message to respect the 50-72 rule.

Please check before I merge: [`pr3902`](https://github.com/Genymobile/scrcpy/commits/pr3902)