@metayan 2023-04-12
> Please check before I merge: [`pr3902`](https://github.com/Genymobile/scrcpy/commits/pr3902)

Checked and it works fine - as intended.

(Under some special conditions there is sometimes still quite a gap between when the device screen is turned off and the `scrcpy` screen appears, but I suspect it would be too complicated to turn off the device screen even later.
The gap is much shorter with this fix, though, so time to let got of some perfectionism...)