@rom1v 2023-04-13
> Under some special conditions there is sometimes still quite a gap between when the device screen is turned off and the scrcpy screen appears

Weird. How long? Can you reproduce a starting delay without turning the screen off? With another encoder (see `scrcpy --list-encoders`)? With another device? With another computer?