@metayan 2023-04-13
Timings __without__ any other OpenGL/Metal-using programs:

### h264 'OMX.qcom.video.encoder.avc'
```
./run x --verbosity=verbose --render-driver=opengl -S \
  --video-codec=h264 --video-encoder='OMX.qcom.video.encoder.avc' \
  2>&1 |ts -i "%.s" |ts -s "%.s"
```
```
0.008024 0.007690 scrcpy 2.0 <https://github.com/Genymobile/scrcpy>
0.008521 0.000254 DEBUG: ADB device found:
0.008567 0.000022 DEBUG:     -->   (usb)  BH901SGS9E                      device  G8441
0.008592 0.000013 DEBUG: Device serial: BH901SGS9E
0.008614 0.000012 DEBUG: Using SCRCPY_SERVER_PATH: x/server/scrcpy-server
0.045004 0.036881 x/server/scrcpy-server: 1 file pushed, 0 skipped. 98.1 MB/s (52576 bytes in 0.001s)
0.397182 0.352163 [server] INFO: Device: Sony G8441 (Android 9)
0.410796 0.013628 [server] DEBUG: Creating encoder by name: 'OMX.qcom.video.encoder.avc'
0.416833 0.006034 [server] WARN: Audio disabled: it is not supported before Android 11
0.416895 0.000059 [server] DEBUG: Audio encoder stopped
0.457307 0.040133 DEBUG: Server connected
0.457471 0.000310 DEBUG: Starting controller thread
0.457512 0.000098 DEBUG: Starting receiver thread
1.411399 0.953957 INFO: Renderer: opengl
1.411442 0.000054 DEBUG: Creating OpenGL Core Profile context
1.432754 0.021267 INFO: OpenGL version: 4.1 NVIDIA-14.0.32 355.11.11.10.10.143
1.432828 0.000081 INFO: Trilinear filtering enabled
1.432855 0.000032 DEBUG: Using SCRCPY_ICON_PATH: app/data/icon.png
1.445209 0.012351 DEBUG: Demuxer 'video': starting thread
1.445266 0.000103 DEBUG: Demuxer 'audio': starting thread
1.445292 0.000056 VERBOSE: input: power mode off
1.445382 0.000046 WARN: Demuxer 'audio': stream explicitly disabled by the device
1.473740 0.028338 INFO: Texture: 720x1280
1.662478 0.188737 [server] INFO: Device screen turned off

2.782003 1.119535 DEBUG: User requested to quit
2.782047 0.000061 DEBUG: quit...
2.782399 0.000369 DEBUG: Demuxer 'video': end of frames
2.782442 0.000047 DEBUG: Receiver stopped
2.796258 0.013803 [server] DEBUG: Controller stopped
3.785394 0.989093 WARN: Killing the server...
3.786020 0.000667 DEBUG: Server disconnected
3.786092 0.000055 DEBUG: Server terminated
```

### h264 'OMX.google.h264.encoder'
```
./run x --verbosity=verbose --render-driver=opengl -S \
  --video-codec=h264 --video-encoder='OMX.google.h264.encoder' \
  2>&1 |ts -i "%.s" |ts -s "%.s"
```
```
0.005748 0.006740 scrcpy 2.0 <https://github.com/Genymobile/scrcpy>
0.006165 0.000273 DEBUG: ADB device found:
0.006194 0.000022 DEBUG:     -->   (usb)  BH901SGS9E                      device  G8441
0.006214 0.000012 DEBUG: Device serial: BH901SGS9E
0.006233 0.000013 DEBUG: Using SCRCPY_SERVER_PATH: x/server/scrcpy-server
0.046630 0.040532 x/server/scrcpy-server: 1 file pushed, 0 skipped. 51.9 MB/s (52576 bytes in 0.001s)
0.411004 0.364611 [server] INFO: Device: Sony G8441 (Android 9)
0.427502 0.016505 [server] DEBUG: Creating encoder by name: 'OMX.google.h264.encoder'
0.442682 0.015125 [server] WARN: Audio disabled: it is not supported before Android 11
0.442738 0.000072 [server] DEBUG: Audio encoder stopped
0.451225 0.008445 DEBUG: Server connected
0.451352 0.000136 DEBUG: Starting controller thread
0.451389 0.000051 DEBUG: Starting receiver thread
1.286342 0.834987 INFO: Renderer: opengl
1.286389 0.000057 DEBUG: Creating OpenGL Core Profile context
1.308164 0.021740 INFO: OpenGL version: 4.1 NVIDIA-14.0.32 355.11.11.10.10.143
1.308239 0.000079 INFO: Trilinear filtering enabled
1.308267 0.000029 DEBUG: Using SCRCPY_ICON_PATH: app/data/icon.png
1.320757 0.012509 DEBUG: Demuxer 'video': starting thread
1.320819 0.000078 DEBUG: Demuxer 'audio': starting thread
1.320845 0.000045 VERBOSE: input: power mode off
1.320870 0.000023 WARN: Demuxer 'audio': stream explicitly disabled by the device
1.346838 0.025961 INFO: Texture: 720x1280
1.540273 0.193398 [server] INFO: Device screen turned off

2.928698 1.388447 DEBUG: User requested to quit
2.928741 0.000054 DEBUG: quit...
2.929107 0.000375 DEBUG: Demuxer 'video': end of frames
2.929152 0.000050 DEBUG: Receiver stopped
2.940539 0.011378 [server] DEBUG: Controller stopped
3.933722 0.993129 WARN: Killing the server...
3.934347 0.000688 DEBUG: Server disconnected
3.934404 0.000047 DEBUG: Server terminated
```

### h265 'OMX.qcom.video.encoder.hevc'
```
./run x --verbosity=verbose --render-driver=opengl -S \
  --video-codec=h265 --video-encoder='OMX.qcom.video.encoder.hevc' \
  2>&1 |ts -i "%.s" |ts -s "%.s"
```
```
0.012964 0.012130 scrcpy 2.0 <https://github.com/Genymobile/scrcpy>
0.013387 0.000280 DEBUG: ADB device found:
0.013432 0.000034 DEBUG:     -->   (usb)  BH901SGS9E                      device  G8441
0.013471 0.000025 DEBUG: Device serial: BH901SGS9E
0.013497 0.000022 DEBUG: Using SCRCPY_SERVER_PATH: x/server/scrcpy-server
0.052338 0.039147 x/server/scrcpy-server: 1 file pushed, 0 skipped. 108.4 MB/s (52576 bytes in 0.000s)
0.404325 0.352074 [server] INFO: Device: Sony G8441 (Android 9)
0.416890 0.012570 [server] DEBUG: Creating encoder by name: 'OMX.qcom.video.encoder.hevc'
0.422464 0.005515 [server] WARN: Audio disabled: it is not supported before Android 11
0.422541 0.000097 [server] DEBUG: Audio encoder stopped
0.452923 0.030414 DEBUG: Server connected
0.452982 0.000092 DEBUG: Starting controller thread
0.453013 0.000029 DEBUG: Starting receiver thread
1.400158 0.947103 INFO: Renderer: opengl
1.400201 0.000057 DEBUG: Creating OpenGL Core Profile context
1.422178 0.021930 INFO: OpenGL version: 4.1 NVIDIA-14.0.32 355.11.11.10.10.143
1.422251 0.000081 INFO: Trilinear filtering enabled
1.422279 0.000027 DEBUG: Using SCRCPY_ICON_PATH: app/data/icon.png
1.435335 0.013071 DEBUG: Demuxer 'video': starting thread
1.435396 0.000101 DEBUG: Demuxer 'audio': starting thread
1.435461 0.000053 VERBOSE: input: power mode off
1.435490 0.000034 WARN: Demuxer 'audio': stream explicitly disabled by the device
1.463124 0.027563 INFO: Texture: 720x1280
1.646464 0.183371 [server] INFO: Device screen turned off

2.664422 1.017992 DEBUG: User requested to quit
2.664466 0.000056 DEBUG: quit...
2.664787 0.000329 DEBUG: Demuxer 'video': end of frames
2.664828 0.000044 DEBUG: Receiver stopped
2.677581 0.012698 [server] DEBUG: Controller stopped
3.665437 0.987851 WARN: Killing the server...
3.666184 0.000743 DEBUG: Server disconnected
3.666252 0.000086 DEBUG: Server terminated
```

This is on macOS Catalina (10.15.7)
MacBook Pro 2.5 GHz Quad-Core Intel Core i7
NVIDIA GeForce GT 750M