@metayan 2023-04-14
> And without `--render-driver=opengl`?

Good point.
Similar delays.

### With lots of windows and tabs in Safari
```
./run x --verbosity=verbose -S \
  --video-codec=h264 --video-encoder='OMX.qcom.video.encoder.avc' \
  2>&1 |ts -i "%.s" |ts -s "%.s"
```
```
0.002456 0.002025 scrcpy 2.0 <https://github.com/Genymobile/scrcpy>
0.003342 0.000415 DEBUG: ADB device found:
0.003376 0.000039 DEBUG:     -->   (usb)  BH901SGS9E                      device  G8441
0.003399 0.000027 DEBUG: Device serial: BH901SGS9E
0.003422 0.000026 DEBUG: Using SCRCPY_SERVER_PATH: x/server/scrcpy-server
0.051011 0.048373 x/server/scrcpy-server: 1 file pushed, 0 skipped. 26.1 MB/s (52576 bytes in 0.002s)
1.058392 1.007360 [server] INFO: Device: Sony G8441 (Android 9)
1.074980 0.016635 [server] DEBUG: Creating encoder by name: 'OMX.qcom.video.encoder.avc'
1.083481 0.008496 [server] WARN: Audio disabled: it is not supported before Android 11
1.083523 0.000043 [server] DEBUG: Audio encoder stopped
1.166413 0.082820 DEBUG: Server connected
1.166496 0.000138 DEBUG: Starting controller thread
1.166555 0.000042 DEBUG: Starting receiver thread
6.195868 5.029260 INFO: Renderer: metal
        [^^^]
6.196058 0.000120 DEBUG: Trilinear filtering disabled (not an OpenGL renderer
6.196143 0.000036 DEBUG: Using SCRCPY_ICON_PATH: app/data/icon.png
6.306953 0.110712 DEBUG: Demuxer 'video': starting thread
6.320481 0.000079 DEBUG: Demuxer 'audio': starting thread
6.321911 0.000030 VERBOSE: input: power mode off
6.331767 0.011775 WARN: Demuxer 'audio': stream explicitly disabled by the device
6.520577 0.202020 [server] INFO: Device screen turned off
7.471523 0.912220 INFO: Texture: 720x1280

9.772552 2.339766 DEBUG: User requested to quit
9.772618 0.000078 DEBUG: quit...
9.773856 0.001174 DEBUG: Demuxer 'video': end of frames
9.773917 0.000117 DEBUG: Receiver stopped
9.783215 0.009319 [server] DEBUG: Controller stopped
10.778078 0.994817 WARN: Killing the server...
10.779054 0.001003 DEBUG: Server disconnected
10.779111 0.000071 DEBUG: Server terminated
```

### With nothing going on
```
./run x --verbosity=verbose -S \
  --video-codec=h264 --video-encoder='OMX.qcom.video.encoder.avc' \
  2>&1 |ts -i "%.s" |ts -s "%.s"
```
```
0.006046 0.005661 scrcpy 2.0 <https://github.com/Genymobile/scrcpy>
0.006576 0.000350 DEBUG: ADB device found:
0.006610 0.000035 DEBUG:     -->   (usb)  BH901SGS9E                      device  G8441
0.006636 0.000022 DEBUG: Device serial: BH901SGS9E
0.006660 0.000021 DEBUG: Using SCRCPY_SERVER_PATH: x/server/scrcpy-server
0.035490 0.029289 x/server/scrcpy-server: 1 file pushed, 0 skipped. 131.8 MB/s (52576 bytes in 0.000s)
0.360263 0.324766 [server] INFO: Device: Sony G8441 (Android 9)
0.377530 0.017270 [server] DEBUG: Creating encoder by name: 'OMX.qcom.video.encoder.avc'
0.380638 0.003086 [server] WARN: Audio disabled: it is not supported before Android 11
0.380726 0.000101 [server] DEBUG: Audio encoder stopped
0.386727 0.006057 DEBUG: Server connected
0.386767 0.000050 DEBUG: Starting controller thread
0.386784 0.000023 DEBUG: Starting receiver thread
1.209370 0.822518 INFO: Renderer: metal
        [^^^]
1.209419 0.000085 DEBUG: Trilinear filtering disabled (not an OpenGL renderer
1.209440 0.000034 DEBUG: Using SCRCPY_ICON_PATH: app/data/icon.png
1.221447 0.011936 DEBUG: Demuxer 'video': starting thread
1.221522 0.000111 DEBUG: Demuxer 'audio': starting thread
1.221539 0.000038 VERBOSE: input: power mode off
1.221558 0.000021 WARN: Demuxer 'audio': stream explicitly disabled by the device
1.263713 0.042146 INFO: Texture: 720x1280
1.436921 0.173181 [server] INFO: Device screen turned off

2.419327 0.982421 DEBUG: User requested to quit
2.419376 0.000055 DEBUG: quit...
2.419738 0.000383 DEBUG: Demuxer 'video': end of frames
2.419784 0.000041 DEBUG: Receiver stopped
2.428461 0.008679 [server] DEBUG: Controller stopped
2.589124 0.160600 [server] DEBUG: Screen streaming stopped
2.589210 0.000089 [server] DEBUG: Device message sender stopped
2.867756 0.278565 DEBUG: Server disconnected
2.867837 0.000092 DEBUG: Server terminated
```

macOS seems to be quite tricky to get to behave when setting up
OpenGL/Metal.

I have now tried some examples both from SDL and GLFW, and the setting
up and removal of the OpenGL context makes the whole system stagger.
Even more during teardown than setup actually. For example, if audio
is playing, it stutters.

So this behaviour is not specific to `scrcpy`.

So far, the only program I have found that behaves well is `mpv`.

Just for comparison - and maybe inspiration - here is an example of
`mpv` playing the first half second of a video (with lots of windows
and tabs in Safari):

```
./build/mpv --end=0.5 --keep-open=no test.mp4 --msg-level=libmpv_render=v,vo=v \
  |ts -i "%.s" |ts -s "%.s" |tee timing
```
```
0.099105 0.097076 [libmpv_render] GL_VERSION='4.1 INTEL-14.7.28'
[Here the OpenGL context is already set up]
0.099539 0.000311 [libmpv_render] Detected desktop OpenGL 4.1.
0.099560 0.000033 [libmpv_render] GL_VENDOR='Intel Inc.'
0.099572 0.000026 [libmpv_render] GL_RENDERER='Intel Iris Pro OpenGL Engine'
0.099584 0.000024 [libmpv_render] GL_SHADING_LANGUAGE_VERSION='4.10'
[...]
0.112607 0.012386  (+) Video --vid=1 (*) (h264 1920x1080 29.970fps)
0.119624 0.006995 [vo/libmpv] Assuming 59.990270 FPS for display sync.
[...]
0.282833 0.050312 [libmpv_render] Window size: 3840x2160 (Borders: l=0 t=0 r=0 b=0)
0.282912 0.000071 [libmpv_render] Video source: 1920x1080 (1:1)
0.282947 0.000016 [libmpv_render] Video display: (0, 0) 1920x1080 -> (0, 0) 3840x2160
[...]
0.746894 0.000089 Exiting... (End of file)
```
The whole setup, playing, and teardown is done in three quarters of a second.
Without any disruption to any other program running.
Hmm.